Update prodcutsController.js products Post route
Removed product key from post route. Added additional table id verification checks that ensure id is not less than 0 when querying. This change was meant to be synced from the 11/8 Merge, but did not get pushed with my other changes as the file had conflicts and was not staged when I thought it was.

 I have manually reviewed the main version of this file to the commit here and am sure these changes are correct